Moving e2e tests into standard location
