+++ e2e-tests/fixtures/permissions/permissions.json
