+++ e2e-tests/images/minio.Dockerfile
