+++ e2e-tests/images/restorer.Dockerfile
